requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence-transformers
PyPDF2
langchain-objectbox

langchain-huggingface
langchain-community
langchain
langserve
fastapi
uvicorn

chromadb
faiss-cpu

langchain-ollama